requirements.txt
flask
python-dotenv
google-generativeai
PyPDF2
gunicorn
